requirements.txt
discord.py~=2.3.2
python-dotenv~=1.0.0
asyncpg~=0.28.0
aiofiles~=23.2.1
aioEasyPillow~=0.0.3